+++ src/action/reserve.ts
@@ -1,7 +1,6 @@
 import * as ActionFactory from '../action';
 import { ActionType } from '../actionType';
 import { AssetTransactionType } from '../assetTransactionType';
-import { OrderType } from '../order';
 import { IAvailableReservationStatusType } from '../reservation';
 import { IReservationFor as IBusReservationFor } from '../reservation/busReservation';
 import { IReservationFor as IEventReservationFor } from '../reservation/event';
@@ -14,27 +13,15 @@ export type IAgent = ActionFactory.IParticipantAsProject;
 export type IReservationFor = (Pick<IBusReservationFor, 'id' | 'typeOf'> | Pick<IEventReservationFor, 'id' | 'typeOf'>) & {
     optimized: boolean;
 };
-// ReservationPackageに拡張(2022-12-22~)
-export interface IReservationPackageAsObject {
+// IReservationPackageAsObject以外を廃止(2023-01-19~)
+export interface IObject {
     reservationFor: IReservationFor;
     reservationNumber: string;
     reservationStatus: IAvailableReservationStatusType;
-    // 不要なので廃止(2023-01-19~)
-    // subReservation?: ISubReservation[];
-    // 不要なので廃止(2023-05-31~)→予約取引から取得に変更
-    // underName?: IUnderName;
     typeOf: ReservationType.ReservationPackage;
 }
-// IReservationPackageAsObject以外を廃止(2023-01-19~)
-// export type IObject = IBusReservation | IEventReservation | IReservationPackageAsObject;
-export type IObject = IReservationPackageAsObject;
 export interface IResult {
     confirmedReservationId?: string;
-}
-export interface IOrderAsReservePurpose {
-    typeOf: OrderType.Order;
-    confirmationNumber?: string;
-    orderNumber?: string;
 }
 export interface IAssetTransactionAsReservePurpose {
     /**