+++ src/action/transfer/return/paymentMethod.ts
@@ -1,11 +1,13 @@
 import * as ActionFactory from '../../../action';
+import { IInvoice } from '../../../invoice';
 import { IReferencedInvoice, ISimpleOrder } from '../../../order';
 import { IAttributes as ISendEmailMessageActionAttributes } from '../../transfer/send/message/email';
 import * as ReturnActionFactory from '../return';
 
 export type IAgent = ActionFactory.IParticipantAsProject;
 export type IRecipient = ActionFactory.IParticipant;
-export type IObject = IReferencedInvoice;
+export type IObject = Pick<IReferencedInvoice, 'accountId' | 'issuedThrough' | 'paymentMethod' | 'paymentMethodId' | 'totalPaymentDue' | 'name' | 'additionalProperty'>
+    & Pick<IInvoice, 'typeOf'>; // typeOfを定義(2024-03-06~)
 export type IResult = any;
 export interface IPotentialActions {
     /**