+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - update @motionpicture/coa-service
 - update @motionpicture/gmo-service
+- 施設コンテンツのcoaInfoを拡張
 
 ### Deprecated