+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - update @motionpicture/coa-service
 - update @motionpicture/gmo-service
 - 施設コンテンツのcoaInfoを拡張
+- 施設コンテンツのlocationを最適化(kanaNameを廃止)
 
 ### Deprecated